Added the dimens and fixed the error causing the error!
The way I resolved the error was through looking at the project structure and changing my SDK & the latest gradel sync. Also, I was being stuck up with an error which said the "primaryDark10" is not related to the "PrimaryDark0". And by changing the code both way, I found out about hybrid sync.